requirements.txt
eel==0.14.0
pyyaml
paramiko